server: move domain types to a separate file
