+++ server/src/domain/mod.rs
@@ -9,3 +9,4 @@ pub mod sql_migrations;
 pub mod sql_opaque_handler;
 pub mod sql_tables;
 pub mod sql_user_backend_handler;
+pub mod types;
